las coordenadas de las actividades salian como texto y rompian la lista, las devuelvo como numero

+++ src/repositories/activity.repository.js
@@ -6,7 +6,7 @@ export async function findMany({ where = [], params = [], orderBy = 'a.date_star
   const rows = await query(
     `SELECT a.id, a.title, a.description, a.modality, a.difficulty, a.kind, a.status,
             a.location,
-            COALESCE(a.lat, p.lat) AS lat, COALESCE(a.lng, p.lng) AS lng,
+            COALESCE(a.lat, p.lat)::float8 AS lat, COALESCE(a.lng, p.lng)::float8 AS lng,
             a.price, a.capacity, a.seats_left,
             a.date_start, a.date_end, a.created_at,
             a.enable_running, a.enable_deposit, a.deposit_percent, a.has_capacity_limit, a.enable_files,
